refactor(mouse): move the test clusters to their post-split subjects (C4)
The F4 decomposition moved the code but left every test in
`mouse/mod.rs`, which is why it read as 1,815 lines against ~307 of
production. Pure relocation: no test logic changed, no renames beyond
what module paths force.

    mod.rs        1815 → 303   (zero tests remain)
    route.rs       525 → 1611  (51 tests)
    selection.rs   398 → 837   (10 tests)

61 `#[test]` before, 61 after.

Five tests filed under the list/chrome/pane selection sections turned
out to assert `route_mouse`'s decision rather than the selection
machinery — `a_covering_pager_outranks_row_selection` and friends check
which sink the router picks, not what the selection does afterwards.
They moved to route.rs with the rest of the routing tests. The section
headings had grouped them by the feature they concern rather than the
function they exercise.

No test needed to stay behind as cross-cutting: every one had a single
clear subject once the routing/selection line was drawn properly.

Paths adjusted where module depth changed (`super::forward::` meant
`mouse::forward::` from mod.rs and is one level further from a child's
test module); imports reconciled per destination.
